Add more graphics on map and organize tiles
