--- lib/Map/Tile/grass.rb
@@ -1,7 +1,0 @@
-require_relative 'tile.rb'
-
-class Grass < Tile
-  def initialize
-    super(description: "You are standing on some grass.")
-  end
-end
