--- lib/Map/Tile/stone.rb
@@ -1,7 +1,0 @@
-require_relative 'tile.rb'
-
-class Stone < Tile
-  def initialize
-    super(description: "You are standing on stone ground.", graphic: '□')
-  end
-end
