--- lib/Map/Tile/wall.rb
@@ -1,7 +1,0 @@
-require_relative 'tile.rb'
-
-class Wall < Tile
-  def initialize
-    super(passable: false)
-  end
-end
